feat: feat: add TaskService as intermediary between controller and database. decouples controller from business logic

+++ src/main/java/taylor/project/projecttracker/Repository/DeveloperRepository.java
@@ -3,5 +3,5 @@ package taylor.project.projecttracker.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import taylor.project.projecttracker.Entity.Developer;
 
-public interface DeveloperRepository extends JpaRepository<Developer, Integer> {
+public interface DeveloperRepository extends JpaRepository<Developer, Long> {
 }
